+++ lib/KBaseSearchEngine/KBaseSearchEngineClient.py
@@ -68,7 +68,7 @@ class KBaseSearchEngine(object):
            parameter "min_date" of Long, parameter "max_date" of Long,
            parameter "min_double" of Double, parameter "max_double" of
            Double, parameter "exclude_subobjects" of type "boolean" (A
-           boolean. 0 = false, other = true.), parameter "lookupInKeys" of
+           boolean. 0 = false, other = true.), parameter "lookup_in_keys" of
            mapping from String to type "MatchValue" (Optional rules of
            defining constraints for values of particular term (keyword).
            Appropriate field depends on type of keyword. For instance in case
@@ -145,7 +145,7 @@ class KBaseSearchEngine(object):
            parameter "min_date" of Long, parameter "max_date" of Long,
            parameter "min_double" of Double, parameter "max_double" of
            Double, parameter "exclude_subobjects" of type "boolean" (A
-           boolean. 0 = false, other = true.), parameter "lookupInKeys" of
+           boolean. 0 = false, other = true.), parameter "lookup_in_keys" of
            mapping from String to type "MatchValue" (Optional rules of
            defining constraints for values of particular term (keyword).
            Appropriate field depends on type of keyword. For instance in case
@@ -201,8 +201,7 @@ class KBaseSearchEngine(object):
            "boolean" (A boolean. 0 = false, other = true.), parameter
            "skip_data" of type "boolean" (A boolean. 0 = false, other =
            true.), parameter "include_highlight" of type "boolean" (A
-           boolean. 0 = false, other = true.), parameter "data_includes" of
-           list of String
+           boolean. 0 = false, other = true.)
         :returns: instance of type "SearchObjectsOutput" (Output results for
            'search_objects' method. 'pagination' and 'sorting_rules' fields
            show actual input for pagination and sorting. total - total number
@@ -225,22 +224,34 @@ class KBaseSearchEngine(object):
            "boolean" (A boolean. 0 = false, other = true.), parameter
            "ascending" of type "boolean" (A boolean. 0 = false, other =
            true.), parameter "objects" of list of type "ObjectData"
-           (Properties of found object including metadata, raw data and
-           keywords. mapping<string, list<string>> highlight - The keys are
-           the field names and the list contains the sections in each field
-           that matched the search query. Fields with no hits will not be
+           (Properties of an object including metadata, raw data and
+           keywords. GUID guid - the object's guid. GUID parent_guid - the
+           guid of the object's parent if the object is a subobject (e.g.
+           features for genomes). object_name - the object's name. timestamp
+           - the creation date for the object in milliseconds since the
+           epoch. parent_data - raw data extracted from the subobject's
+           parent object. The data contents will vary from object to object.
+           Null if the object is not a subobject. data - raw data extracted
+           from the object. The data contents will vary from object to
+           object. key_props - keyword properties of the object. These fields
+           have been extracted from the object and possibly transformed based
+           on the search specification for the object. The contents will vary
+           from object to object. mapping<string, string> object_props -
+           general properties for all objects. This mapping contains the keys
+           'creator', 'copied', 'module', 'method', 'module_ver', and
+           'commit' - respectively the user that originally created the
+           object, the user that copied this incarnation of the object, and
+           the module and method used to create the object and their version
+           and version control commit hash. Not all keys may be present; if
+           not their values were not available in the search data.
+           mapping<string, list<string>> highlight - The keys are the field
+           names and the list contains the sections in each field that
+           matched the search query. Fields with no hits will not be
            available. Short fields that matched are shown in their entirety.
-           Longer fields are shown as snippets preceded or followed by "...".
-           mapping<string, string> object_props - general properties for all
-           objects. This mapping contains the keys 'creator', 'copied',
-           'module', 'method', 'module_ver', and 'commit' - respectively the
-           user that originally created the object, the user that copied this
-           incarnation of the object, and the module and method used to
-           create the object and their version and version control commit
-           hash. Not all keys may be present; if not their values were not
-           available in the search data.) -> structure: parameter "guid" of
-           type "GUID" (Global user identificator. It has structure like
-           this: <data-source-code>:<full-reference>[:<sub-type>/<sub-id>]),
+           Longer fields are shown as snippets preceded or followed by
+           "...".) -> structure: parameter "guid" of type "GUID" (Global user
+           identificator. It has structure like this:
+           <data-source-code>:<full-reference>[:<sub-type>/<sub-id>]),
            parameter "parent_guid" of type "GUID" (Global user identificator.
            It has structure like this:
            <data-source-code>:<full-reference>[:<sub-type>/<sub-id>]),
@@ -294,28 +305,38 @@ class KBaseSearchEngine(object):
            true.), parameter "skip_keys" of type "boolean" (A boolean. 0 =
            false, other = true.), parameter "skip_data" of type "boolean" (A
            boolean. 0 = false, other = true.), parameter "include_highlight"
-           of type "boolean" (A boolean. 0 = false, other = true.), parameter
-           "data_includes" of list of String
+           of type "boolean" (A boolean. 0 = false, other = true.)
         :returns: instance of type "GetObjectsOutput" (Output results of
            get_objects method. mapping<access_group_id, narrative_info>
            access_group_narrative_info - information about the workspaces in
            which the objects in the results reside. This data only applies to
            workspace objects.) -> structure: parameter "objects" of list of
-           type "ObjectData" (Properties of found object including metadata,
-           raw data and keywords. mapping<string, list<string>> highlight -
-           The keys are the field names and the list contains the sections in
-           each field that matched the search query. Fields with no hits will
-           not be available. Short fields that matched are shown in their
-           entirety. Longer fields are shown as snippets preceded or followed
-           by "...". mapping<string, string> object_props - general
-           properties for all objects. This mapping contains the keys
+           type "ObjectData" (Properties of an object including metadata, raw
+           data and keywords. GUID guid - the object's guid. GUID parent_guid
+           - the guid of the object's parent if the object is a subobject
+           (e.g. features for genomes). object_name - the object's name.
+           timestamp - the creation date for the object in milliseconds since
+           the epoch. parent_data - raw data extracted from the subobject's
+           parent object. The data contents will vary from object to object.
+           Null if the object is not a subobject. data - raw data extracted
+           from the object. The data contents will vary from object to
+           object. key_props - keyword properties of the object. These fields
+           have been extracted from the object and possibly transformed based
+           on the search specification for the object. The contents will vary
+           from object to object. mapping<string, string> object_props -
+           general properties for all objects. This mapping contains the keys
            'creator', 'copied', 'module', 'method', 'module_ver', and
            'commit' - respectively the user that originally created the
            object, the user that copied this incarnation of the object, and
            the module and method used to create the object and their version
            and version control commit hash. Not all keys may be present; if
-           not their values were not available in the search data.) ->
-           structure: parameter "guid" of type "GUID" (Global user
+           not their values were not available in the search data.
+           mapping<string, list<string>> highlight - The keys are the field
+           names and the list contains the sections in each field that
+           matched the search query. Fields with no hits will not be
+           available. Short fields that matched are shown in their entirety.
+           Longer fields are shown as snippets preceded or followed by
+           "...".) -> structure: parameter "guid" of type "GUID" (Global user
            identificator. It has structure like this:
            <data-source-code>:<full-reference>[:<sub-type>/<sub-id>]),
            parameter "parent_guid" of type "GUID" (Global user identificator.