+++ RELEASE_NOTES.md
@@ -6,6 +6,8 @@ Search Service MKII release notes
 
 * Adds more information to the error thrown when encountering an unexpected type in a recursive
   index
+* Removed `data_includes` from the API, as it was unimplemented.
+* Changed `lookupInKeys` to lookup\_in_keys to be consistent with other fields.
 
 0.1.0
 -----